Extract HTTPBinForUncleanShutdown
