+++ Tests/AsyncHTTPClientTests/HTTPClientTestUtils.swift
@@ -931,103 +931,6 @@ final class ConnectionsCountHandler: ChannelInboundHandler {
     }
 }
 
-internal class HttpBinForSSLUncleanShutdown {
-    let group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
-    let serverChannel: Channel
-
-    var port: Int {
-        return Int(self.serverChannel.localAddress!.port!)
-    }
-
-    init(channelPromise: EventLoopPromise<Channel>? = nil) {
-        self.serverChannel = try! ServerBootstrap(group: self.group)
-            .serverChannelOption(ChannelOptions.socket(SocketOptionLevel(SOL_SOCKET), SO_REUSEADDR), value: 1)
-            .childChannelOption(ChannelOptions.socket(IPPROTO_TCP, TCP_NODELAY), value: 1)
-            .childChannelInitializer { channel in
-                let requestDecoder = HTTPRequestDecoder()
-                return channel.pipeline.addHandler(ByteToMessageHandler(requestDecoder)).flatMap {
-                    let configuration = TLSConfiguration.makeServerConfiguration(certificateChain: [.certificate(try! NIOSSLCertificate(bytes: Array(cert.utf8), format: .pem))],
-                                                                                 privateKey: .privateKey(try! NIOSSLPrivateKey(bytes: Array(key.utf8), format: .pem)))
-                    let context = try! NIOSSLContext(configuration: configuration)
-                    return channel.pipeline.addHandler(NIOSSLServerHandler(context: context), name: "NIOSSLServerHandler", position: .first).flatMap {
-                        channel.pipeline.addHandler(HttpBinForSSLUncleanShutdownHandler(channelPromise: channelPromise))
-                    }
-                }
-            }.bind(host: "127.0.0.1", port: 0).wait()
-    }
-
-    func shutdown() {
-        try! self.group.syncShutdownGracefully()
-    }
-}
-
-internal final class HttpBinForSSLUncleanShutdownHandler: ChannelInboundHandler {
-    typealias InboundIn = HTTPServerRequestPart
-    typealias OutboundOut = ByteBuffer
-
-    let channelPromise: EventLoopPromise<Channel>?
-
-    init(channelPromise: EventLoopPromise<Channel>? = nil) {
-        self.channelPromise = channelPromise
-    }
-
-    func channelRead(context: ChannelHandlerContext, data: NIOAny) {
-        switch self.unwrapInboundIn(data) {
-        case .head(let req):
-            self.channelPromise?.succeed(context.channel)
-
-            let response: String?
-            switch req.uri {
-            case "/nocontentlength":
-                response = """
-                HTTP/1.1 200 OK\r\n\
-                Connection: close\r\n\
-                \r\n\
-                foo
-                """
-            case "/nocontent":
-                response = """
-                HTTP/1.1 204 OK\r\n\
-                Connection: close\r\n\
-                \r\n
-                """
-            case "/noresponse":
-                response = nil
-            case "/wrongcontentlength":
-                response = """
-                HTTP/1.1 200 OK\r\n\
-                Connection: close\r\n\
-                Content-Length: 6\r\n\
-                \r\n\
-                foo
-                """
-            default:
-                response = """
-                HTTP/1.1 404 OK\r\n\
-                Connection: close\r\n\
-                Content-Length: 9\r\n\
-                \r\n\
-                Not Found
-                """
-            }
-
-            if let response = response {
-                var buffer = context.channel.allocator.buffer(capacity: response.count)
-                buffer.writeString(response)
-                context.writeAndFlush(self.wrapOutboundOut(buffer), promise: nil)
-            }
-
-            context.channel.pipeline.removeHandler(name: "NIOSSLServerHandler").whenSuccess {
-                context.close(promise: nil)
-            }
-        case .body:
-            ()
-        case .end:
-            ()
-        }
-    }
-}
-
 internal final class CloseWithoutClosingServerHandler: ChannelInboundHandler {
     typealias InboundIn = HTTPServerRequestPart
     typealias OutboundOut = HTTPServerResponsePart
